Separate Loader into its own component
